% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kfolds.R
\name{blendModels}
\alias{blendModels}
\title{Compute model blending ratio}
\usage{
blendModels(model.results.soc2, model.results.soc3, convergence.iteration)
}
\value{

}
\description{
DETAILED DESCRIPTION
}
