% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{computeCIs}
\alias{computeCIs}
\title{Calculate confidence intervals for predictions (missing observations only)}
\usage{
computeCIs(blended.results, confidence.level = 0.95)
}
\value{

}
\description{
DETAILED DESCRIPTION
}
