% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kfolds.R
\name{computeConvergence}
\alias{computeConvergence}
\title{Compute convergence iteration}
\usage{
computeConvergence(test.fold.data, verbose = TRUE)
}
\arguments{
\item{test.fold.data}{Test fold predictions (output of getTestFoldData())}

\item{verbose}{Should messages be printed; default is TRUE (print messages)}
}
\value{
Convergence iteration of k-folds cross validation modeling
}
\description{
DETAILED DESCRIPTION Recall that Iteration 0 is simply the output of smart
guessing (no modeling involved).
}
