% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_impute.R
\name{computeMeanPredictions}
\alias{computeMeanPredictions}
\title{Get predictions (one per simulation) and overall mean prediction for a given
iteration}
\usage{
computeMeanPredictions(simulations, iteration, ors.data.sims)
}
\value{

}
\description{
DETAILED DESCRIPTION
}
