% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{computeStdEVdiff}
\alias{computeStdEVdiff}
\title{Calculate differences in standardized EVs between two jobs, and generate plot}
\usage{
computeStdEVdiff(blended.results, jobA, jobB, print.plot = TRUE)
}
\arguments{
\item{print.plot}{}
}
\value{

}
\description{
DETAILED DESCRIPTION
}
