% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{correlationPlot}
\alias{correlationPlot}
\title{Compute correlation of ELEs, and plot relevant heatmap (grouped by SOC2 code)}
\usage{
correlationPlot(blended.results, plot.dim = 6.5, print.plot = TRUE)
}
\arguments{
\item{blended.results}{Blended predictions from imputation models, calculated
at convergence iterations and blending proportions computed by
\code{\link[=computeBlendingRatio]{computeBlendingRatio()}} (output of \code{\link[=blendImputations]{blendImputations()}})}

\item{plot.dim}{Length of each side of the plot (in inches); default is 6.5}

\item{print.plot}{Should plot file (.png) be generated; default is TRUE
(create plot file)}
}
\value{
A list of length four, containing a data frame of expected values, a
correlation matrix, and two plot objects (one where the individual
occupations are labeled in addition to the SOC2 group labels, and one where
they are not); optionally produces a PDF of the heatmap
}
\description{
Produces a heatmap displaying Pearson correlation between expected values
(ELEs) by occupation. The heatmap is arranged according to SOC2 code, so that
similar occupations are grouped together.
}
\seealso{
\code{\link[=blendImputations]{blendImputations()}}

\code{\link[=computeEVs]{computeEVs()}}

\code{\link[stats:cor]{stats::cor()}}
}
