% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{errorsAndBounding}
\alias{errorsAndBounding}
\title{Assign errors to known mean estimates, and bounds for all estimates (known
and unknown)}
\usage{
errorsAndBounding(blsE, blsMnGC)
}
\arguments{
\item{blsE}{Errors associated with mean estimates (output of getErrors())}

\item{blsMnGC}{N-1 group completed data (output of fillNminusOneGroups())}
}
\value{
Data with relevant errors and bounds assigned to each record
}
\description{
The errors are collected alongside the mean estimates, and can allow for
empirical bounding of the data during modeling. For records lacking a mean
estimate/error, we can still bound the missing estimate using the remaining
value in an occupational group. For example, if there are three records in an
occupational group, of which two estimates are known and sum to X\%, the upper
bound of the missing estimate is 100\%-X\%, and the lower bound is 0\%.
}
