% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_simulation.R
\name{estBetaParams}
\alias{estBetaParams}
\title{Estimate beta distribution parameters}
\usage{
estBetaParams(mu, var)
}
\arguments{
\item{mu}{Mean of distribution}

\item{var}{Variance (error) of distribution}
}
\value{
Beta distribution parameters
}
\description{
Generates parameters (alpha and beta) to define a beta distribution based on
the input data (mean and variance).
}
