% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{fillNminusOneGroups}
\alias{fillNminusOneGroups}
\title{Calculate mean estimates for N-1 groups}
\usage{
fillNminusOneGroups(blsMnC)
}
\arguments{
\item{blsMnC}{Mean estimates (output of getMeanEstimates())}
}
\value{
N-1 group completed data
}
\description{
Some of the occupational groups are only missing a single mean estimate.
E.g., binary occupational groups (with requirement levels 'YES'/'NO') that
have one estimate available can be filled in based on the known data.
Similarly, if an occupational group has 5 levels, and 4 of them have known
estimates, then the final one can be calculated. This is because the total
value of any occupational group must be 100\%.
}
