% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{getErrors}
\alias{getErrors}
\title{Get standard error estimates}
\usage{
getErrors(orsC)
}
\arguments{
\item{orsC}{Completed, corrected data (output of \code{\link[=dataCorrections]{dataCorrections()}})}
}
\value{
Standard error estimates only
}
\description{
Select only the standard error estimates from the data.
}
\seealso{
\code{\link[=dataCorrections]{dataCorrections()}}

\code{\link[=getMeanEstimates]{getMeanEstimates()}}
}
