% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{getMeanPredictions}
\alias{getMeanPredictions}
\title{Get convergence predictions (one per simulation) and overall mean prediction}
\usage{
getMeanPredictions(simulations, iteration, ors.data.sims)
}
\value{

}
\description{
DETAILED DESCRIPTION
}
