% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{getORS}
\alias{getORS}
\title{Get raw ORS data from BLS}
\usage{
getORS(orsLink = "https://www.bls.gov/ors/xlsx/2018_excel_output.xlsx")
}
\arguments{
\item{orsLink}{pointing to excel data file at BLS;
default is https://www.bls.gov/ors/xlsx/2018_excel_output.xlsx}
}
\value{
Raw ORS data from BLS website
}
\description{
Retrieve raw Occupational Requirements Survey (ORS) data from the Bureau of
Labor Statistics' (BLS) website.
}
