% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kfolds.R
\name{getTestFoldData}
\alias{getTestFoldData}
\title{Collate predictions for all 10 test folds, for each iteration}
\usage{
getTestFoldData(model.iterations)
}
\arguments{
\item{model.iterations}{Modeling results (output of iterateModelKFCVwSG(),
specifically output$model.iterations)}
}
\value{
Collated predictions of each test fold, by iteration
}
\description{
Takes the results of a k-folds CV run and collates all the test fold data.
In the KFCV analysis, each of 10 folds is treated as the test fold, and a
model is generated and iterated over each of these test folds. The
predictions resulting from each test fold are combined for each iteration,
such that there is a complete predicted dataset (all records) obtained from
each iteration.
}
