% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_traintest.R
\name{getTestSetData}
\alias{getTestSetData}
\title{Collate predictions for each iteration for the train-test approach}
\usage{
getTestSetData(model.results)
}
\arguments{
\item{model.results}{Modeling results (output of iterateModelTT())}
}
\value{
Predictions from the test set, by iteration
}
\description{
Takes the results of a train-test run and collates all the test set data.
The model is generated and iterated over the test set, producing predictions
for about 10\% of the data. These predictions are extracted and organized by
iteration. Recall that Iteration 0 is simply the naive guess (no modeling
involved).
}
