% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_impute.R
\name{iterateModel}
\alias{iterateModel}
\title{Impute missing estimates, with Smart Guessing}
\usage{
iterateModel(
  ors.data.sims,
  n.iter,
  weight.step,
  mdl.d = 14,
  mdl.n = 200,
  mdl.e = 0.6,
  sg.soc.code
)
}
\arguments{
\item{ors.data.sims}{Original data augmented with relevant predictors, i.e.
all records, including both known and missing estimates, as well as simulated
data (output of \code{\link[=computeSimulations]{computeSimulations()}})}

\item{n.iter}{Number of times to iterate/adjust the model}

\item{weight.step}{Increment by which to increase modeling weight of missing
data with each iteration}

\item{mdl.d}{Tree model maximum depth; default is 14}

\item{mdl.n}{Tree model rounds; default is 200}

\item{mdl.e}{Tree model eta; default is 0.6}

\item{sg.soc.code}{SOC code to use for smart guessing, either "upSOC2",
"upSOC3", or "upSOC4"}
}
\value{
A list containing the results of iterative modeling, for each
simulation
}
\description{
Function to impute missing estimates in an iterative fashion. The first
prediction (Iteration 0) is the result of the smart guessing procedure. All
subsequent iterations rely on XGBoost to produce preliminary predictions.
These predictions are then adjusted to adhere to boundary conditions on the
data (all estimates must be in the range [0,1], and the sum of all estimates
within an occupational group must be <=1).
}
\seealso{
\code{\link[=computeSimulations]{computeSimulations()}}

\code{\link[=smartGuess]{smartGuess()}}

\code{\link[xgboost:xgb.train]{xgboost::xgboost()}}

\code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel()}}

\code{\link[parallel:makeCluster]{parallel::makeCluster()}}
}
