% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kfolds.R
\name{iterateModelKFCV}
\alias{iterateModelKFCV}
\title{K-folds CV, with Smart Guessing}
\usage{
iterateModelKFCV(
  ors.data,
  n.iter,
  weight.step,
  mdl.d = 14,
  mdl.n = 200,
  mdl.e = 0.6,
  fold.list = NULL,
  sg.soc.code
)
}
\arguments{
\item{ors.data}{Original data augmented with relevant predictors, i.e. all
records, including both known and missing estimates (output of
getPredictors(), or computeSimulations())}

\item{n.iter}{Number of times to iterate/adjust the model}

\item{weight.step}{Increment by which to increase modeling weight of test
fold data with each iteration}

\item{mdl.d}{Tree model maximum depth; default is 14}

\item{mdl.n}{Tree model rounds; default is 200}

\item{mdl.e}{Tree model eta; default is 0.6}

\item{fold.list}{A list of preset folds, if these have to be fixed across
runs; default is to generate a new set of folds}

\item{sg.soc.code}{SOC code to use for smart guessing, either "upSOC2",
"upSOC3", or "upSOC4"}
}
\value{
A list of length two, containing a list of hold out indices for each
test fold, and the results of iterative modeling
}
\description{
Function to do k-folds cross validation in an iterative fashion. The first
prediction (Iteration 0) is the result of the smart guessing procedure. All
subsequent iterations rely on XGBoost to produce preliminary predictions.
These predictions are then adjusted to adhere to boundary conditions on the
data (all estimates must be in the range \link{0,1}, and the sum of all estimates
within an occupational group must be <=1).
}
