% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kfolds.R
\name{iterateModelKFCVwSG}
\alias{iterateModelKFCVwSG}
\title{K-folds CV with Smart Guessing}
\usage{
iterateModelKFCVwSG(
  original.ors.full,
  n.iter,
  initial.guess.weight,
  prediction.contribution,
  guess.contribution,
  weight.step,
  mdl.d,
  mdl.n,
  mdl.e,
  fold.list = NULL,
  sg.soc.code
)
}
\arguments{
\item{original.ors.full}{Original data (all records, including both known and
missing estimates)}

\item{n.iter}{Number of times to adjust/iterate the model}

\item{sg.soc.code}{SOC code to use for smart guessing, either "upSOC2",
"upSOC3", or "upSOC4"}
}
\value{

}
\description{
DETAILED DESCRIPTION lord have mercy on my soul
}
