% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kfolds.R
\name{plotTestFolds}
\alias{plotTestFolds}
\title{Plot predicted vs. actual values}
\usage{
plotTestFolds(
  test.fold.data,
  plot.alpha,
  n.rounds,
  reqcat.palette = c("black", "yellow", "cyan"),
  line.color = c("red"),
  titles = gsub("([a-z])([0-9])", "\\\\1 \\\\2", colnames(test.fold.data))
)
}
\arguments{
\item{test.fold.data}{Test fold predictions (output of getTestFoldData())}

\item{plot.alpha}{Alpha (transparency) value to use for points}

\item{n.rounds}{How many iterations are to be plotted}

\item{reqcat.palette}{A character vector specifying colors to use for each
requirement category; defaults are black (ENV), yellow (PHY), and cyan (EDU)}

\item{line.color}{A string specifying color to use for the line y=x; default
is red}

\item{titles}{A character vector specifying the titles of each subplot;
default is the column names of test.fold.data}
}
\value{
A plot object
}
\description{
DETAILED DESCRIPTION
}
