% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{predictCIs}
\alias{predictCIs}
\title{Calculate confidence intervals for predictions (missing observations only)}
\usage{
predictCIs(model.summary.average, confidence.level = 0.95)
}
\value{

}
\description{
DETAILED DESCRIPTION
}
