% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predictors.data}
\alias{predictors.data}
\title{Predictor mapping}
\format{
A data frame with 299 rows and 7 variables:
\describe{
\item{data_element_text}{original data field describing the requirement}
\item{data_type_text}{original data field describing the requirement level}
\item{additive_group}{original data field specifying the additive group}
\item{Requirement}{new, refined data field describing the requirement}
\item{Frequency}{new data field describing the requirement level (1 of 2)}
\item{Intensity}{new data field describing the requirement level (2 of 2)}
\item{Requirement_category}{new data field describing the requirement
category}
}
}
\source{
See \url{https://www.bls.gov/ors/information-for-survey-participants/pdf/occupational-requirements-survey-collection-manual-082020.pdf}
for the guidance used to arrive at Frequency/Intensity/Requirement_category
mappings.
}
\usage{
predictors.data
}
\description{
A data frame describing the refining of textual (categorical) information in
the ORS dataset, along with the mapping of this refined information to three
new predictors: Frequency (numerical), Intensity (numerical), and
Requirement_category (categorical). This mapping is used to define the
predictor columns for the data in preparation for modeling.
}
\seealso{
\code{\link[=getPredictors]{getPredictors()}}
}
\keyword{datasets}
