% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kfolds.R
\name{smartGuessKFCV}
\alias{smartGuessKFCV}
\title{Smart Guessing for k-folds CV}
\usage{
smartGuessKFCV(
  ors.data.for.sg,
  ors.data,
  wt.low = 0,
  wt.mid = 0.5,
  wt.high = 0.5
)
}
\arguments{
\item{ors.data.for.sg}{Data with "missing" estimates to populate with smart
guesses}

\item{ors.data}{Original data augmented with relevant predictors, i.e. all
records, including both known and missing estimates (output of
getPredictors(), or getSimulations())}

\item{wt.low}{Model weight to assign to low-confidence smart guesses; default
is 0}

\item{wt.mid}{Model weight to assign to mid-confidence smart guesses; default
is 0.5}

\item{wt.high}{Model weight to assign to high-confidence smart guesses;
default is 0.5}
}
\value{
Input data frame, with missing values filled in by smart guesses
}
\description{
DETAILED DESCRIPTION lord have mercy on my soul
}
