% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{standardizeEVs}
\alias{standardizeEVs}
\title{Standardize expected values, and generate plot}
\usage{
standardizeEVs(blended.results, print.plot = TRUE)
}
\arguments{
\item{print.plot}{}
}
\value{

}
\description{
DETAILED DESCRIPTION
}
