% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_simulation.R
\name{stdErrMC}
\alias{stdErrMC}
\title{Draw sample from specified beta distribution}
\usage{
stdErrMC(mu, stdev, n = 10)
}
\arguments{
\item{mu}{Mean of distribution}

\item{stdev}{Variance (error) of distribution}

\item{n}{Number of values to draw; default is 10}
}
\value{
A sample of size n
}
\description{
This function generates a sample of n values derived from a beta distribution
specified by a mean and associated variance (error). The beta distribution is
generated using estBetaParams(). The default sample size is 10.
}
