% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{transformORS}
\alias{transformORS}
\title{Transform raw ORS data}
\usage{
transformORS()
}
\value{

}
\description{
The data retrieved from BLS website needs to be transformed into the format
expected by all subsequent preprocessing (and other) functions in this
package. This function accomplishes this task.
}
