% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_impute.R
\name{blendImputations}
\alias{blendImputations}
\title{Blend model results according to calculated proportions}
\usage{
blendImputations(
  model.results.soc2,
  model.results.soc3,
  conv.iter.soc2,
  conv.iter.soc3,
  soc2.prop,
  soc3.prop,
  write.files = FALSE
)
}
\arguments{
\item{model.results.soc2}{}

\item{model.results.soc3}{}

\item{conv.iter.soc2}{}

\item{conv.iter.soc3}{}

\item{soc2.prop}{}

\item{soc3.prop}{}

\item{write.files}{}

\item{ors.data.sims}{}
}
\value{

}
\description{
Details
}
