% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kfolds.R
\name{computeConvergence}
\alias{computeConvergence}
\title{Compute convergence iteration}
\usage{
computeConvergence(test.fold.data, verbose = TRUE, show.plot = TRUE)
}
\arguments{
\item{test.fold.data}{Test fold predictions (output of getTestFoldData())}

\item{verbose}{Should messages be printed; default is TRUE (print messages)}

\item{show.plot}{Should plot of RMSE by iteration be displayed; default is
TRUE (display plot)}
}
\value{
Convergence iteration of k-folds cross validation modeling
}
\description{
For a set of iterative k-folds modeling results, it is necessary to determine
the convergence point of the model. This function computes this using the
RMSE of iterated predictions vs. the actual known values (this can be
accomplished because the k-folds procedure relies on known estimates only,
and simulates missing estimates from the known set).
}
\details{
We defined convergence as when the difference between the RMSE of consecutive
iterations was < 0.001. If this convergence criteria is not met, the function
simply returns the final iteration (and a message stating that convergence
was not reached). Recall that Iteration 0 is simply the output of smart
guessing (no modeling involved).
}
