% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_impute.R
\name{computeModelUncertainty}
\alias{computeModelUncertainty}
\title{Approximate uncertainty due to models}
\usage{
computeModelUncertainty(
  model.results.soc2,
  model.results.soc3,
  conv.iter.soc2,
  conv.iter.soc3,
  soc2.prop,
  soc3.prop
)
}
\arguments{
\item{model.results.soc2}{}

\item{model.results.soc3}{}

\item{conv.iter.soc2}{}

\item{conv.iter.soc3}{}

\item{soc2.prop}{}

\item{soc3.prop}{}
}
\value{

}
\description{
DETAILED DESCRIPTION
}
