% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{computeOverlap}
\alias{computeOverlap}
\title{Compute overlap score between two jobs, and generate plot}
\usage{
computeOverlap(blended.results, jobA, jobB, print.plot = TRUE)
}
\arguments{
\item{print.plot}{}
}
\value{

}
\description{
DETAILED DESCRIPTION
}
