% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{correlationPlot}
\alias{correlationPlot}
\title{Compute correlation of EVs, and plot relevant heatmap (grouped by SOC2 code)}
\usage{
correlationPlot(blended.results, print.plot = TRUE)
}
\arguments{
\item{print.plot}{}
}
\value{

}
\description{
DETAILED DESCRIPTION
}
