% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{dataCorrections}
\alias{dataCorrections}
\title{Make corrections to data}
\usage{
dataCorrections(blsC)
}
\arguments{
\item{blsC}{Completed data (output of fillMissingObservations(), or
otherMissingObservations())}
}
\value{
Data, updated with corrections
}
\description{
During our analysis process we noticed some corrections/adjustments that
needed to be made:
}
\details{
(1) Requirement 'Hearing requirements: Pass a hearing test' was split between
'Hearing requirements: Pass a hearing test' and
'Hearing requirements: Pass a hearing Test'; these are standardized to the
former.

(2) Remove observations with 'FULLY MITIGATED' in the data_type_text field.

(3) Levels for the synthetic additive group (requirement) 78, ('Sitting or
standing/walking') are set as 'SIT' and 'STAND' under data_type_text.

(4) Ensure that data is formatted as a data frame.
}
