% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{getErrors}
\alias{getErrors}
\title{Get standard error estimates}
\usage{
getErrors(blsC)
}
\arguments{
\item{blsC}{Completed, corrected data (output of dataCorrections())}
}
\value{
Standard error estimates only
}
\description{
Select only the standard error estimates from the data.
}
