% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{getMeanEstimates}
\alias{getMeanEstimates}
\title{Get mean estimates}
\usage{
getMeanEstimates(blsC)
}
\arguments{
\item{blsC}{Completed, corrected data (output of dataCorrections())}
}
\value{
Mean estimates only
}
\description{
Select only the mean estimates from the data.
}
