% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{getPredictors}
\alias{getPredictors}
\title{Populate predictors based on data mapping}
\usage{
getPredictors(orsMnGCe)
}
\arguments{
\item{orsMnGCe}{Data with errors and bounds (output of errorsAndBounding())}
}
\value{
Data augmented with relevant predictors
}
\description{
(1) The data_element_text field was restructured into a new field,
Requirement, and added to the data.
}
\details{
(2) The data_element_text and data_type_text fields have somewhat overlapping
roles in the original data.  Moreover their structure makes for a large
increase in the dimensions of the data. We devised a way to map these
two fields to numeric fields (Frequency and Intensity), leading to a
significant reduction in dimension. This mapping is applied here, and the new
fields are added to the data.

(3) Also added to the data are fields for SOC2, SOC3, and SOC4 codes, as well
as a field for the relevant Requirement Category.

(4) Generate an indicator column that differentiates between known estimates
and missing estimates.
}
