% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kfolds.R
\name{iterateModelKFCVwSG}
\alias{iterateModelKFCVwSG}
\title{K-folds CV with Smart Guessing}
\usage{
iterateModelKFCVwSG(
  original.bls.full,
  n.iter,
  initial.guess.weight,
  prediction.contribution,
  guess.contribution,
  weight.step,
  mdl.d,
  mdl.n,
  mdl.e,
  fold.list = NULL,
  sg.soc.code
)
}
\value{

}
\description{
DETAILED DESCRIPTION lord have mercy on my soul
}
