% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{otherMissingObservations}
\alias{otherMissingObservations}
\title{Generate additional (optional) missing observations}
\usage{
otherMissingObservations(blsC)
}
\arguments{
\item{blsC}{Completed data (output of fillMissingObservations())}
}
\value{
Completed data, with requested requirement/levels added
}
\description{
SSA analysts requested the addition of a few other missing observations,
based on requirement/level, listed below:
}
\details{
Humidity: CONSTANTLY
Heavy vibrations: CONSTANTLY, FREQUENTLY, SELDOM
Extreme cold: CONSTANTLY, FREQUENTLY
Climbing ladders, ropes, or scaffolds: CONSTANTLY, FREQUENTLY
Climbing ramps or stairs (work-related): CONSTANTLY, FREQUENTLY
Stooping is missing: CONSTANTLY
Kneeling is missing: CONSTANTLY, FREQUENTLY
Crawling is missing: CONSTANTLY, FREQUENTLY
}
