% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kfolds.R
\name{plotTestFolds}
\alias{plotTestFolds}
\title{Plot predicted vs. actual values}
\usage{
plotTestFolds(
  test.fold.data,
  plot.alpha = 0.1,
  n.iter = (ncol(test.fold.data) - 4),
  reqcat.palette = c("palevioletred1", "black", "yellow", "cyan"),
  line.color = c("red"),
  titles = gsub("([a-z])([0-9])", "\\\\1 \\\\2", colnames(test.fold.data))
)
}
\arguments{
\item{test.fold.data}{Test fold predictions (output of getTestFoldData())}

\item{plot.alpha}{Alpha (transparency) to use for points; default is 0.1}

\item{n.iter}{How many iterations are to be plotted; default is to plot all
iterations, including Iteration 0 (smart guessing)}

\item{reqcat.palette}{A character vector specifying colors to use for each
requirement category; defaults are pink (COG), black (ENV), yellow (PHY), and
cyan (EDU)}

\item{line.color}{A string specifying color to use for the line y=x; default
is red}

\item{titles}{A character vector specifying the titles of each subplot;
default is the column names of test.fold.data}
}
\value{
Produces a plot (.png file) and returns a list of plot objects, with
one object for each iteration
}
\description{
DETAILED DESCRIPTION Recall that Iteration 0 is simply the output of smart
guessing (no modeling involved). Note that n.iter=10 will plot Iterations
0-10, for a total of 11 iterations.
}
