% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kfolds.R
\name{plotTestFolds}
\alias{plotTestFolds}
\title{Plot predicted vs. actual values from k-folds CV results}
\usage{
plotTestFolds(
  test.fold.data,
  plot.alpha = 0.1,
  n.iter = (ncol(test.fold.data) - 4),
  reqcat.palette = c("palevioletred1", "black", "yellow", "cyan"),
  line.color = c("red"),
  titles = gsub("([a-z])([0-9])", "\\\\1 \\\\2", colnames(test.fold.data)),
  print.plot = TRUE
)
}
\arguments{
\item{test.fold.data}{Test fold predictions (output of \code{\link[=getTestFoldData]{getTestFoldData()}})}

\item{plot.alpha}{Alpha (transparency) to use for points; default is 0.1}

\item{n.iter}{How many iterations are to be plotted; default is to plot all
iterations, including Iteration 0 (smart guessing)}

\item{reqcat.palette}{A character vector specifying colors to use for each
requirement category; defaults are pink (COG), black (ENV), yellow (PHY), and
cyan (EDU)}

\item{line.color}{A string specifying color to use for the line y=x; default
is red}

\item{titles}{A character vector specifying the titles of each subplot;
default is the column names of test.fold.data}

\item{print.plot}{Should plot file (.png) be generated; default is TRUE
(create plot file)}
}
\value{
Returns a list of plot objects, with one object for each iteration;
optionally produces a plot (.png file) of all iterations
}
\description{
For each iteration, plot the predicted values vs. actual values of
observations in test folds. Recall that Iteration 0 is simply the output of
smart guessing (no modeling involved). Note that n.iter=10 will plot
Iterations 0-10, for a total of 11 iterations.
}
\seealso{
\code{\link[=iterateModelKFCV]{iterateModelKFCV()}}

\code{\link[=getTestFoldData]{getTestFoldData()}}
}
