% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{setDefaultModelingWeights}
\alias{setDefaultModelingWeights}
\title{Assign default modeling weights}
\usage{
setDefaultModelingWeights(ors.data, unknown.weight = 0.5)
}
\arguments{
\item{ors.data}{Data augmented with relevant predictors (output of
getPredictors())}

\item{unknown.weight}{Modeling weight to assign to unknown estimates; default
is 0.5}
}
\value{
Data augmented with relevant predictors and default modeling weights
}
\description{
In the modeling stage, data points can be weighted differently. This function
assigns default weights to records based on the presence/absence of a mean
estimate.
}
