% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kfolds.R
\name{smartGuessKFCV}
\alias{smartGuessKFCV}
\title{Smart Guessing for k-folds CV}
\usage{
smartGuessKFCV(
  ors.data,
  original.ors.full,
  sim.no = "value",
  wt.low = 0,
  wt.mid = 0.5,
  wt.high = 0.5
)
}
\arguments{
\item{ors.data}{Data with "missing" estimates to populate with smart guesses}

\item{original.ors.full}{Original data (all records, including both known and
missing estimates)}

\item{sim.no}{Is the smart guessing being run on a simulation, or the
original value; default is original value}

\item{wt.low}{Model weight to assign to low-confidence smart guesses; default
is 0}

\item{wt.mid}{Model weight to assign to mid-confidence smart guesses; default
is 0.5}

\item{wt.high}{Model weight to assign to high-confidence smart guesses;
default is 0.5}
}
\value{
Original data frame, with missing values filled in by smart guesses
}
\description{
DETAILED DESCRIPTION lord have mercy on my soul
}
